requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
transformers>=4.35.0
torch>=2.0.0
streamlit>=1.28.0

